ThreeEqual.java feedback

+++ ThreeEqual.java
@@ -3,9 +3,10 @@ public class ThreeEqual {
         if (args.length<3){
             System.out.println("not enough variables");
         }else{
-            double a = Double.parseDouble(args[0]);
-            double b = Double.parseDouble(args[1]);
-            double c = Double.parseDouble(args[2]);
+            // read the README carefully for specifications on input/output
+            int a = Integer.parseInt(args[0]);
+            int b = Integer.parseInt(args[1]);
+            int c = Integer.parseInt(args[2]);
             if (a == b && b==c){
                 System.out.println("equal");
             }else{
